pyproject.toml
[project]
name = "pydantic-ai-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "debugpy>=1.8.20",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.61.0",
    "python-dotenv>=1.2.1",
]
